homepage has been added logo's manipulation has been added
