+++ vis_helpers/vis_utils.py
@@ -19,35 +19,58 @@ def trim_spectra(df):
 
 
 @st.cache
-def show_logo():
+def show_logo(width, padding, margin):
     with open('logos/logo.png', 'rb') as f:
         data = f.read()
-
+    
+    padding_top, padding_right, padding_bottom, padding_left = padding
+    margin_top, margin_right, margin_bottom, margin_left = margin
+    
     bin_str = base64.b64encode(data).decode()
     html_code = f'''
                     <img src="data:image/png;base64,{bin_str}"
                     style="
-                         margin: auto;
-                         margin-top:-30px;
-                         width: 65%;
-                         padding:0px 6px 20px 25%;
-                         "/>
+                     margin: auto;
+                     width: {width}%;
+                     margin-top: {margin_top}px;
+                     margin-right: {margin_right}px;
+                     margin-bottom: {margin_bottom}px;
+                     margin-left: {margin_left}%;
+                     padding-top: {margin_top}px;
+                     padding-right: {padding_right}px;
+                     padding-bottom: {padding_bottom}px;
+                     padding-left: {padding_left}%;
+                     "/>
                 '''
     return html_code
 
 
 @st.cache
-def show_sersitivis_logo():
+def show_sersitivis_logo(width, padding, margin):
+    padding_top, padding_right, padding_bottom, padding_left = padding
+    margin_top, margin_right, margin_bottom, margin_left = margin
+    
     link = 'http://sersitive.eu'
-
+    
     with open('logos/sersitivis_no_background.png', 'rb') as f:
         data = f.read()
-
+    
     bin_str = base64.b64encode(data).decode()
     html_code = f'''
         <a href="{link}" target = _blank>
             <img src="data:image/png;base64,{bin_str}"
-            style="padding:0px 6px 5px 0px; 20px; height:80px"/>
+            style="
+                 margin: auto;
+                 width: {width}%;
+                 margin-top: {margin_top}px;
+                 margin-right: {margin_right}px;
+                 margin-bottom: {margin_bottom}px;
+                 margin-left: {margin_left}%;
+                 padding-top: {margin_top}px;
+                 padding-right: {padding_right}px;
+                 padding-bottom: {padding_bottom}px;
+                 padding-left: {padding_left}%;
+                 "/>
         </a>'''
 
     return html_code
